requirements.txt
streamlit==1.32.0
requests
pandas
pillow
python-dotenv
